+++ Changelog.md
@@ -18,6 +18,8 @@
 - Fix: Error when unbinding the last mine/sentry from a Mine/Sentry Remote Access Tool
 - Fix: Placing an Inventory Scanner facing a disguised scanner does not update the placed scanner's disguise
 - Fix: The Cage Trap and Password-protected Furnace/Smoker/Blast Furnace are black instead of see-through
+- Fix: Breaking an Inventory Scanner/Laser or one of its field does not break all fields, if the respective config setting is lower than the amount of fields present
+- Fix: Breaking an Inventory Scanner Fields with the inventoryScannerRange config setting set to >=3 does not break and reinstate all fields properly
 
 --------------------------Changelog for v1.9.4 of SecurityCraft--------------------------